requirements.txt
pillow==8.3.2
psutil==5.8.0
PyGetWindow==0.0.9
matplotlib==3.4.3
seaborn==0.11.2
pandas==1.3.3
scikit-learn==0.24.2
schedule==1.1.0
requests==2.26.0
customtkinter
CTkMessagebox
